Implemented RecvPrepare and SendCommit predicates.
Implemented the receiving of a Prepare message from a peer
and the sending of a Commit message when the necessary quorum
of Prepares is collected.

Minor changes:
Moved c.WF() in v.WF() and passing the constants to the variables WF.
Added getters for the maximum faulty nodes tolerated in Constants and
the agreement quorum size.
Fixed passing of v instead of v' in NextStep predicate.