Initial support for signatures modeling.
We provide an abstraction to verify signatures
by consulting with the trusted network entity in
the system, which keeps track of all messages and
can detect when a host has tried to impersonate an
other peer host.

In this MR we have split the messages in 2 "frames".
The outer one contains the sender of the message and
the inner frame - the payload. In this way when the
Distributed system selects a host for the step it
passes its ID to the network and the network can
check whether this host has filled correctly its ID
in the "sender" field of the outer frame. This way
the network keeps track of all correctly signed
messages and can answer queries by the `msgOps`
binding variable, where we have added an additional
parameter - a set called `signatureChecks`. In each
host we can fill this set with the messages for which
we want a verification that they really originate
from the entity reflected in "sender".