Received ViewChangeMsg-s storage in Replicas implementation.
We store the received View Change Msg-s in a nested map,
first we split the messages by the View, then in the
nested map by sender.
Added Predicate that indicates whether out View is active.
If we have sent a View Change Msg for the View we are in,
and we have less that 2F+1 ViewChangeMsgs for this View,
it is inactive.
If we have more than F+1 ViewChangeMsgs for a higher View
than ours, our View is inactive.
When we add the NewViewMsg from the primary we will add it
to this predicate accordingly.