Removed unnecessary restrictions in predicates.
In HaveSufficientVCMsgsToMoveTo we can move to any
View beyond our own as long as we have at least F+1
VC messages for a higher view without breaking safety.
It also does not hurt liveness, because we will
eventually reach the correct view, even if we perform
multiple steps.

In ViewIsActive the check whether we have F+1 View Change
messages for a higher View is redundant. If we have them,
we will eventually move to a higher View, which will
initially be inactive before we collect the necessary
information we need to activate it.