Added New View message.
We have added the action for the Primary of a View to
select a set of 2F+1 View Change messages and to pack
them inside a NewViewMsg. The Primary can then send
this message to the peers so that Replicas will be able
to go to the next phase of the View Change process.