We don't allow replicas to directly impersonate each other.
In this version of the Network we add a check whether the sender has
filled correctly its ID in the message. This way we don't allow direct
impersonation of peers by malicious replicas, which in the real world
is provided by the signatures that messages carry.
Even though this check the sender can still put non existent (not
previously sent by peers) messages in the New View message or the
View Change message, which contain in themselves quorums of messages
generated from peer replicas, and we need to check those for
validity in replica.i.dfy